added auth and updated homepage
