+++ src/app/(auth)/layout.tsx
@@ -1,0 +1,11 @@
+export default function AuthLayout({
+    children,
+  }: Readonly<{
+    children: React.ReactNode;
+  }>) {
+    return (
+      <main className="min-h-screen flex items-center justify-center">
+        {children}
+      </main>
+    );
+  }
