+++ src/app/(dashboard)/page.tsx
@@ -1,0 +1,109 @@
+"use client";
+
+import {
+  Calendar,
+  CalendarCurrentDate,
+  CalendarDayView,
+  CalendarMonthView,
+  CalendarNextTrigger,
+  CalendarPrevTrigger,
+  CalendarTodayTrigger,
+  CalendarViewTrigger,
+  CalendarWeekView,
+  CalendarYearView,
+} from "@/components/ui/full-calendar";
+import { addHours } from "date-fns";
+import { ChevronLeft, ChevronRight } from "lucide-react";
+import { useAuth } from "../context/AuthContext";
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+
+export default function Page() {
+  const { isAuthenticated } = useAuth();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!isAuthenticated) {
+      router.push("/login");
+    }
+  }, [isAuthenticated, router]);
+
+  if (!isAuthenticated) {
+    return null;
+  }
+
+  return (
+    <Calendar
+      events={[
+        {
+          id: "1",
+          start: new Date(),
+          end: addHours(new Date(), 2),
+          title: "Asset A",
+          color: "pink",
+        },
+        {
+          id: "2",
+          start: new Date(),
+          end: addHours(new Date(), 4),
+          title: "Asset B",
+          color: "blue",
+        },
+      ]}
+    >
+      <div className="h-dvh p-14 flex flex-col">
+        <div className="flex px-6 items-center gap-2 mb-6">
+          <CalendarViewTrigger
+            className="aria-[current=true]:bg-accent"
+            view="day"
+          >
+            Day
+          </CalendarViewTrigger>
+          <CalendarViewTrigger
+            view="week"
+            className="aria-[current=true]:bg-accent"
+          >
+            Week
+          </CalendarViewTrigger>
+          <CalendarViewTrigger
+            view="month"
+            className="aria-[current=true]:bg-accent"
+          >
+            Month
+          </CalendarViewTrigger>
+          <CalendarViewTrigger
+            view="year"
+            className="aria-[current=true]:bg-accent"
+          >
+            Year
+          </CalendarViewTrigger>
+
+          <span className="flex-1" />
+
+          <CalendarCurrentDate />
+
+          <CalendarPrevTrigger>
+            <ChevronLeft size={20} />
+            <span className="sr-only">Previous</span>
+          </CalendarPrevTrigger>
+
+          <CalendarTodayTrigger>Today</CalendarTodayTrigger>
+
+          <CalendarNextTrigger>
+            <ChevronRight size={20} />
+            <span className="sr-only">Next</span>
+          </CalendarNextTrigger>
+
+          {/* <ModeToggle /> */}
+        </div>
+
+        <div className="flex-1 px-6 overflow-hidden">
+          <CalendarDayView />
+          <CalendarWeekView />
+          <CalendarMonthView />
+          <CalendarYearView />
+        </div>
+      </div>
+    </Calendar>
+  );
+}
