+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import SideNav from "@/components/SideNav";
+import { AuthProvider } from "./context/AuthContext";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -14,8 +14,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Sitespace",
+  description: "Your all in one site scheduling app",
 };
 
 export default function RootLayout({
@@ -28,28 +28,9 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased bg-orange-50`}
       >
-        {/* Mobile SideNav (only visible on mobile) */}
-        <div className="md:hidden">
-          <SideNav />
-        </div>
-        
-        {/* Desktop layout (hidden on mobile) */}
-        <div className="hidden md:flex h-screen">
-          <div className="w-16 flex-shrink-0 z-50 relative">
-            <SideNav />
-          </div>
-          
-          <div className="flex-1">
-            {children}
-          </div>
-        </div>
-        
-        {/* Mobile content (only visible on mobile) */}
-        <main className="md:hidden min-h-screen">
-          <div className="pt-16">
-            {children}
-          </div>
-        </main>
+        <AuthProvider>
+          {children}
+        </AuthProvider>
       </body>
     </html>
   );
