+++ src/app/unauthorized/page.tsx
@@ -1,0 +1,20 @@
+import Link from 'next/link';
+
+export default function Unauthorized() {
+  return (
+    <div className="flex min-h-screen items-center justify-center">
+      <div className="text-center">
+        <h1 className="text-6xl font-extrabold text-orange-600 mb-4">Unauthorized Access</h1>
+        <p className="text-gray-600 mb-8">
+          You don't have permission to access this page.
+        </p>
+        <Link 
+          href="/"
+          className="inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
+        >
+          Return to Home
+        </Link>
+      </div>
+    </div>
+  );
+}
